requirements.txt
Flask==3.1.1
python-dotenv==1.0.1
httpx==0.28.1